pipeline: neo: Fix core dump after camera stop on SIGINT
Core dump can be observed when pipeline camera stream is interrupted
by SIGINT, because of incorrect handling of cancelled buffers
reported to the pipeline by the stopped video devices.

Add the necessary handling of the framebuffer by immediately
completing all buffers associated to the same frame, and completing
the request.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>