pipeline: neo: Add support for raw stream mode
Raw stream mode implementation is added to the pipeline and
supports two use cases.
First use case is raw-only mode where there is a unique stream
enabled for the camera and the stream type is raw. Neither ISP
nor IPA involved in that mode if operations, raw buffers
acquired from CSI/ISI are simply reported to application.
Second use case correspond to stream mode concurrent with a
stream that requires debayering. In that case ISP and IPA
are used as usual, as it would with a regular stream using
ISP for debayering.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>