ipa: mojom: neo: Add SensorConfig return value to init()
SensorConfig structure is introduced and  added to the list of
IPA init() function return values.
It aims at being a container for the sensor-specific
configurations that are known by IPA and that need to be shared
with the pipeline handler.

Initial usage of this structure is to permit definition of the
DelayedControls parameters, needed to configure the instance
operated by the pipeline.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>