ipa: neo: Create NXP CameraSensorHelper library for extension
The CameraSensorHelper from libipa is subclassed in a local library,
in a similar way it has been done for rpi, in order to add NXP
extensions without changing the core implementation.

Initial intent is to abstract sensor-specific controls for the
camera whose internal control granularity mandates usage of custom
controls instead of using the standard V4L2 definitions.
That is for instance needed for Omnivision OX03C10 control of
exposure and gain that require more granularity that available from
the generic V4L2_CID_ANALOGUE_GAIN and V4L2_CID_EXPOSURE.

Thus helpers are added in order to abstract in CameraSensorHelper
subclass the handling of ControlLists and ControlMapInfo
related to those controls.

An additional dependency to camera sensors controls exists in the
DelayedControls instance operated by the pipeline. It has to be
configured with a list of controls that will be updated by IPA
during pipeline streaming, and their associated delay and priority.
Dedicated helper is also provisioned to define a custom control
list and associated parameters.
IPA retrieves the list during init() and passes it back to the
pipeline handler so that it can configures its DelayedControls
instance accordingly.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>