ipa: neo: Add HdrDecomp algorithm
HdrDecomp algorithm aims at configuring the HDR decompression
blocks of the NEO ISP, to be setup when sensor uses compression.

Decompression parameters (kneepoints, coefficients, offsets...)
are defined in the calibration file associated to the sensor.
In case HdrDecomp entry is not populated, then the ISP block is
not configured by the algorithm.

Two blocks instances exist in the ISP and can be configured
independently in the configuration file.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>