ipa: neo: Remove algos raw mode
Pipeline raw mode supports 2 variants: the raw-only mode that
captures only Bayer frames, and the raw-mode concurrent with
frame decoding that also captures decoded frames after
demosaicing.

During raw-only mode, IPA is not involved by pipeline operation.
For concurrent raw-mode with frame decoding, pipeline interacts
with ISP in the same way it does when raw frames are not
captured.

Thus, for NEO pipeline there is no need to have a specific
handling in IPA for raw capture in the pipeline.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>