ipa: neo: agc: Adapt algo with NEO ISP params and statistics
This implementation is based on the RkISP1 and IPU3 AGC/AEC
mean-based control algorithms.

The ISP measurements used for AGC algorithm are based
on the Statistics Histograms.
Those measurements are used to estimate the luminance and the brightness.
A gain is then computed to achieve a target value.

The algorithm is designed to provide the computed AGC/AEC controls
to the sensor.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>