libcamera: v4l2_subdevice: Add stream support to setFormat()
The ioctl VIDIOC_SUBDEV_S_FMT has been completed to support
streams with the addition of relevant member into the
v4l2_subdev_format structure.

This change adds a new setFormat() method signature that
also includes a stream parameter in addition to the pad to
to enable configuration of streams.

Version of the v4l2-subdev.h header file in libcamera
already has support for stream in its definition of the
v4l2_subdev_format structure.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>