pipeline: neo: isi: Add explicit ISI channel allocator
Current channel allocator does not give the option to specify
the channel identifier to be allocated and picks the first
one found available.

For pipeline configuration based on platform config file, it
is needed to explicitly select an ISI channel. A new variant
of the allocator method is added, taking as a parameter the
desired channel index.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>