pipeline: neo: isi: Remove channel subdevice configuration
Responsability of the media graph subdevices format configuration
being under complete control of the pipeline, this includes ISI
crossbar the and channels subdevices.

Remove the odd channel subdevice format configuration from
ISIPipe::configure() so that it is configured along with other
subdevices by the pipeline.

Capture video device pixel format configuration remains in
configure() as it is still ISI-specific.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>