pipeline: neo: Revamp camera topology discovery and configuration
This change is to better support automatic detection of the
camera media device topology and add the option to use a
configuration file to describe complex topologies.

PipelineConfig class is created to provide pipeline with:
- the routing definition for relevant media device entities
- per-camera explicit graph definition, listing from the sensor
to its capture video channel the list of links and streams
used.

Above definitions are aggregated by the PipelineConfig either
through automatic discovery of the media device, or using a
platform configuration file.

A camera may output multiple streams for main exposure (input0),
short exposure (input1) and embedded data.
Those streams are bundled in a CameraInfo object, that pipeline
can query from PipelineConfig.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>