pipeline: neo: Synchronize buffer count between pipeline and ISI
Constant kBufferCount represents the maximum number of in-flight
Requests that application may push to the pipeline. It translates
into the number of intermediate buffers to provision to serve
those requests.

Separate definitions currently exist in pipeline and ISI device.
For consistency and to simplify the system configuration, keep
only the pipeline definition and propagate it to the ISI device.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>