pipeline: neo: Increase pipeline buffer count to 6
In some use cases with a longer latency to recycle the application
buffers e.g. using GStreamer pipeline with display on Wayland, the
current number of buffers (4) is a bit short and leads to spurious
loss of frames. Reason is that the application buffers are not recycled
fast enough to have a new request pending in the pipeline bundled
to an associated buffer to serve the incoming camera frame.

Increase the buffer count from 4 to 6 for now. Ultimately, that
value should be tunable from the configuration file.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>