pipeline: neo: Revamp ISI buffer allocation scheme to be same as ISP
Current ISI internal buffer allocation scheme consists in having a pool
of buffers automatically allocated by every pipe at the time it is
started. ISIPipe class provides an interface to allocate from and release
to this pool of buffers. The pool of buffers is automatically released
when the pipe is stopped.

ISP allocation scheme is a bit different: it exposes an explicit
interface to allocate buffers between configuration and start, and
an interface to deallocate buffers after being stopped.
Benefit is that buffer pool is available before devices are started,
which is required e.g. to be able to pass shared buffers to IPA before
streaming is started.

This change updates the ISI allocation scheme to match the one of the
ISP. Also, the handling of ISI internal buffers is now bundled to the
NxpNeoFrame life cycle.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>