MA-22237 Fix the crash issue when usb camera connected on imx95
Issue occurs when the usb camera and default camera
connected at the same time, when bootUp, camera service will
always reboot due to the null pointer in populateEntities,
this will cause the failure to build pipeline.

In this case, a UVC device was used to populate the
related entities, pads and links, which was not reasonable.

Ensure the media device was not a "uvcvideo"
before createDevice to fix the issue.

Change-Id: I656a542e83449df287d34a4e358ae2e26e784f17
Signed-off-by: Anle Pan <anle.pan@nxp.com>