ipa: neo: Add support for top lines embedded data
Raw buffers output from ISI device nodes are mapped in the IPA.
When a raw buffer is received by the pipeline, IPA
fillParamsBuffers() interface is invoked to retrieve the params
buffer to be used by the ISP for the decoding.
During that same call, raw buffer identifier is passed to IPA.

A sensor can declare a number of top data lines via CameraHelper
configuration. In that case, CameraHelper implementation shall
also implement the parser that will extract from the embedded
data the sensor register values and format them in a set of
standard controls to be usable by IPA.

After the decoding of the raw frame by the ISP, the IPA
processStatsBuffer() interface is invoked to process the
ISP statistics by the 3A algorithms. At that time, sensor
control values corresponding to that frame are required.
In absence of embedded data, sensor control history maintained
by the pipeline is used. In case embedded data is present,
it is used instead, as considered more reliable.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>