neo: pipeline: Revamp pipeline frontend configuration
Configuration of the pipeline frontend is done at camera stream
configuration time when the pipeline controls a single camera, or
at pipeline creation time in case of multi camera. Reason is
that with multicamera, different sensors may have streams sharing
the same pad of a given entity, which is not compatible with
a format setup at configuration time.

This change simplifies and factorizes the 2 variants of the
frontend configuration, using the same code path.
Sensor format (raw mbus code and size) is nowadays computed
once at validate() time, and stored in the CameraConfiguration
object, to be reused at configure() time.
Orientation requested in the CameraConfiguration is hopefully
correctly handled now.

Note that `frontend` terminology has been adopted for the media
controller device hosting the sensor, CSI and ISI entities in
replacement of `backend` name that was previously used.
That is for consistency with RPi naming that uses `backend` name
for ISP and `frontend` name for the sensor and CSI entities.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>