ipa: neo: Fix metadata parsing with active raw stream
The parsing of top lines of embbeded data by the IPA requires that
the raw buffer can be mapped through mapBuffer() before the stream
is started.
When raw stream is active, raw buffers used by the pipelines are
not internal buffers allocated on the frontend video device, but
buffers allocated coming from the application. Thus, raw buffers
are not known in advance and are not mapped in the IPA.
This change adds correct handling for that case in the IPA, by
skipping embedded data parsing in that particular case.
Previous implementation was trapping in exception, trying to
access the mapped address of an unkown buffer.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>