ipa: neo: agc: Add histogram scale configuration from yaml file
For the AGC algorithm, the histogram scaling parameter can be configurable
from the yaml file.
If this is not configured from yaml file, the default parameter value is used.

For ox03c10, the histogram scaling parameter is configured to 128 to
cover the HCG+LCG HDR exposures ranging up to 15bits values.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>