ipa: neo: camhelper: mx95mbcam: Define the setAGC function for OX03C10
The controlListSetAGC() is implemented according to the HDR3 combined
image output from the OX03C10.
The AGC exposure and gain from AEC algorithm are distributed among the
3 captures: HCG, LCG and VS.

The distribution is based on the total exposure ratio between:
- Long (HCG) to Short (DCG) ratio set to 32 (5 bits shift)
- Long (HCG) to Very Short (VS) ratio set to 1024 (10 bits shift)

This implies a minimal HCG gain set to the ratio HCG/LCG value.
This is to guarantee a ratio between the HCG and LCG gains considering that
the minimal LCG gain is 1.0.
This HCG minimal analog gain approximates 4.5, based on
- an HCG/LCG ratio of 32 and
- a gain conversion ratio of 7.32

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>