ipa: neo: Merge CameraHelper delayed controls and meta data helpers
Current implementation of CameraHelper class exposes two methods to
report camera-specific attributes:
- CameraHelper::delayedControlParams() describes the camera controls
updated by IPA and their associated configuration delay in the sensor.
- CameraHelper::embeddedParams() reports the parameters related to
the embedded data output by the sensor, if any.

Additional camera parameters are expected to be added to the list of
sensor attributes to be reported to IPA.

In order to avoid creating a new method any time a new attribute is
added:
- CameraHelper::Attributes structure is defined to group all the
necessary sensor attributes defined in CameraHelper.
- CameraHelper::attributes() getter function is added to get a pointer
to the attribute structure populated by the CameraHelper instance.

IPA calls to those replaced functions are updated accordingly to
make use of the new one.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>