pipeline: neo: Enable isp media controller device links
The Linux kernel driver for Neo ISP now implements mutable links
to conditionally enable video devices depending on the intended
usage.

Output (capture) devices frame and ir can be individually enabled
to configure the ISP to output image on the frame and/or Ir
devices. Usage of Ir pad is restricted to RGBIr sensors.

Input1 (output) device is enabled for HDR merge within the ISP
which is not supported yet.

Params and stats metadata devices are enabled unconditionally for
IPA operation.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>