ipa: mojom: neo: Add rgbIr indication to SensorConfig
The rgbIr field is added to the SensorConfig structure that is populated
by CameraHelper and returned to pipeline handler by init() call.
Purpose is to let the pipeline handler know that the camera has a RGBIr
CFA, so that the Ir capture path of the ISP can be enabled on a separate
stream.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>