pipeline: neo: Use rgbIr indication from the IPA
Pipeline handler needs to know if the camera has RGBIr CFA to decide
if the Ir capture path of the ISP may be used as a dedicated stream.
Implementation is using sensorIsRgbIr() to get that information, but
so far its return value is hardcoded to false meaning no support.

Now that RGBIr indication comes from IPA, update sensorIsRgbIr() to
use the actual value for that camera.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>