pipeline: imx8-isi: Add optional formatter subdevice
This patch adds support of the formatter subdevice. The formatter subdevice
is considerred optional, and is configured only if it is detected during
the 'match' sequence. Indeed i.MX9 SOCs family integrates such element,
while it was not present in previous i.MX8 SOCs family. However imx8-isi
pipeline is used in both cases.

Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>