pipeline: neo: Crop the top embedded lines of the raw frame fed to ISP
Sensor top lines embedded data are metadata associated to the frame,
not pixel data. Therefore, top lines should do not have to be decoded
by the ISP so that they are also not in the decoded frame buffer.

This change adds awareness of the top lines embedded data into the
streams configuration related functions generateConfiguration() and
validate(): streams decoded by ISP have their size adjusted to
remove the embedded data lines. Raw stream is unchanged because
it does not go through ISP.

Now that ISP driver supports cropping at the input video devices with
the SELECTION API, its configuration is updated in order to crop the
top lines embedded data when present for the sensor in use.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>