pipeline: neo: Fix erroneous frame loss warning
Pipeline handler maintains a sequence counter to track that frame
sequence number coming from the front-end device is contiguous.
That is used to detect camera frame loss.

Frame sequence is to be reset anytime a camera session completes
or restarts, otherwise an erroneous discontinuity will be reported
when a new session begins.

Add the missing frame sequence reset.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>