ipa: neo: Mute noisy log Number of registers adjusted
Log 'Number of registers adjusted' is output to notify that number of
registers passed to the meta data parser has been increased to be a
multiple of 4 which is a requirement to match the frame definition.

Information is relevant when integrating a new sensor, but is too
verbose for normal operation of the pipeline.
Switch this log to Debug priority.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>