ipa: neo: drc: Add support for the global gain configuration
The global gain is applied at the output of the global tonemapping step
where all entries of the Global LUT are multiplied by this gain.
It is used to compensate for the dark scene of an HDR sensor.

The global gain can be configured in yaml file using the "gbl-gain" attribute
defining an u16 value with 8.8 format.
If no configuration is set, the default value set by kGblGain is used.

The global gain is configured to 4096 for ox03c10 to compensate
for the dark scene of this sensor.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>