ipa: neo: Set controls with exposure and gain from IPA context
The exposure and gain values used to set controls are extracted
from the IPA context instead of the Frame context.

This change allows to reduce the delay to set controls by 1 frame.
Indeed, from the algorithms, the exposure and gain values are updated
in the Frame context at frame N+1 whereas the same values are updated
in the IPA context at frame N.

This change has also an effect for the initial setControls(0) call
performed from the IPA start() function.
Indeed, the initial settings from this call are now set to sensor configuration
setting (performed from the algorithm configure() function), versus 0 values
coming from the initial values from the Frame Context.
This prevent the sensor to be set with minimal gain and exposure values
for the first frames for which the AGC algorithm computes very high
exposure and gain value to control. This was observed with the first frames
to be overexposed.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>