ipa: neo: Fix IR Compression table
IR Compression table allows the definition of ranges for
input pixel values, that are individualy remapped using
a range-specific gain and offset.

if (pv < points[N])
    opv = (pv - offsets[N]) * ratios[N] + newpoints[N]

with:
- pv: input pixel value
- opv: output pixel value

If no points[] entry fulfills the range condition for a
pixel value, last entry of offsets/newpoints/ratios is
used as the default case. That is when no points[] value
has matched the condition (pv < points[N]).

Comment added in the rgbir algorithm to precise the case.
Ox05b1s calibration file is updated accordingly.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>