ipa: neo: rgbir: Add Head Color configuration
For RGBIr use case, the Head Color unit of the ISP can not be
configured by the ISP driver as it can not derive the RGBIr
pattern from the input video device format.

This change add support for Head Color configuration as part
of the rgbir control algorithm.

The ox05b1s calibration file is updated accordingly.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>