pipeline: neo: Change StreamLink from tuple to structure
A StreamLink identifies a stream over a media link between two Media
Controller entities. Thus, it is bundling a MediaLink that identifies
source and sink pads and associated entities, and a given stream that
has a dedicated identifier on source and sink pads.

the current StreamLink declaration uses a tuple to embed the MediaLink
along with the stream identifier on the source and on the sink pads.
For clarity, replace the tuple by a structure with some explicit member
names for those 3 elements.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>