pipeline: neo: Add multi camera topology detection
Pipeline handler has to configure the media controller graph
used by a camera. To do so, it has to know about the entities,
links and streams involved in the graph. With that information
it can configure the necessary stream formats and entities
routing.

Current implementation supports 2 graph definition modes:
- A file-based configuration where every links and streams are
explicitly defined as well as the necessary routes.
- An automatic discovery mode, that does not require any
file configuration but limited to basic topologies and single
camera detection.

This change implements a more elaborate automatic discovery mode
where the graph from the sensor to the ISI video nodes is
searched recursively. Updated implementation adds support for
multicamera and multistream topologies detection.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>