ipa: neo: Use the embedded data (ox03c10) from previous frame
This is a temporary workaround to compensate the ox03c10 embedded data
not aligned with the frame sensor setting.
Indeed the exposure and gain setting of the frame is matching
the embedded data from the previous frame.
The control list of current frame is set with embedded data
from previous frame.

This patch prevents from the AGC to produce large variation in the
control values when a frame is applied with sensor controls coming
from outside of the AGC algorithm (such as the manual setting or the
initial setting).

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>