neo: ipa: Set ox03c10 exposure and gain control delays to 3 (versus 2)
The ox03c10 sensor controls were set with a delay of 2 with a reference to
the embedded data being updated in the frame.
However due to the embedded data updated 1 frame earlier than
the frame sensor setting, the ox03c10 control delays should be increased by 1 frame.

When embedded data are disabled, this patch prevents from the AGC to produce
large variation in the control values when a frame is applied with sensor
controls coming from outside of the AGC algorithm (such as the manual
setting or the initial setting).

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>