neo: ipa: Set os08a20 exposure and gain control delays to 3
The os08a20 sensor exposure and gain controls should be set
with a delay of 3 frames.
With this value, the ISP statistics (such as the mean histogram value)
of each frame is reflecting the sensor data.

This patch prevents from the AGC to produce large variation in the control
values when a frame is applied with sensor controls coming from outside
of the AGC algorithm (such as the manual setting or the initial setting).

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>