pipeline: neo: Add ISI pipe chaining support
A single ISI pipe can handle image size with a width limited
to 2048 pixels. That size can be extended to 4096 pixels when
chaining the pipe with the next adjacent pipe. That is done
transparently by the ISI Linux driver.
This has to be taken into account when allocating ISI pipe
index, to make sure a pipe is not used as being chained by
the driver for a large image size on the precedent channel.

As a consequence, this change implements two variants of ISI
channel reservation:
1) ISIDevice::reservePipeBySize(): request the reservation of
ISI pipe(s) according to a given size. Index of the first
reserved pipe(s) is reported, accounting for the possible
chaining of the adjacent pipe.
2) ISIDevice::reservePipeByIndex(): this reservation variant is
aimed at being used for manual reservation via config file.
First index is passed by the caller, ISI device class make the
necessary reservation of the requested pipes.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>