ipa: neo: Skip control setting for frame 0
Skip control setting for frame 0 for which the frame context
doesn't have a relevant configuration for the exposure and the gain.
Indeed the frame context is not initialized at startup.

This workaround prevents some frames from flashing at startup.
This effect can be addressed later by configuring some startup
frames to be hidden.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>