libcamera: ipc_unixsocket: Share stdin and stdout with IPA proxy
When IPA is running in isolated mode, at IPA process creation time
all the inherited file descriptors are closed in the new process
after the fork, with the exception of the file descriptor relevant
to the IPC peer Unix socket.

In order to enable the IPA logging in the console, add stdout and
stderr to the list of file descriptors to be shared with the
isolated process.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>